refactor: remove Toast components imports and insert it into provider
