+++ dashboard/components/feedback-widget/FeedbackWidget.tsx
@@ -459,7 +459,6 @@ const useFeedbackWidget = (defaultState: boolean = false) => {
             </form>
           </div>
         </Modal>
-        {toast && <Toast {...toast} dismissToast={dismissToast} />}
       </>
     );
   };
