+++ dashboard/components/inventory/components/view/InventoryViewHeader.tsx
@@ -1,4 +1,4 @@
-import { ToastProps } from '@components/toast/ToastProvider';
+import { ToastProps } from '@components/toast/Toast';
 import { NextRouter } from 'next/router';
 import { FormEvent, useState } from 'react';
 import Button from '../../../button/Button';
