+++ dashboard/components/inventory/hooks/useInventory/helpers/infiniteScrollFilteredList.ts
@@ -1,4 +1,4 @@
-import { ToastProps } from '@components/toast/ToastProvider';
+import { ToastProps } from '@components/toast/Toast';
 import { SetStateAction } from 'react';
 import settingsService from '../../../../../services/settingsService';
 import { InventoryFilterData, InventoryItem } from '../types/useInventoryTypes';
